+++ src/app/lib/ui/dial-text.ts
@@ -11,7 +11,7 @@ units: "<string>" dsisplay units,
 @Component({
 	selector: 'ap-dial-text',
 	template: `
-            <div class="dial-text">
+            <div class="dial-text mat-app-background">
                 <div class="dial-text-title">{{title}}</div>
                 <div class="dial-text-value">{{value}}</div>
                 <div class="dial-text-units">{{units}}</div>