+++ src/app/lib/ui/pip.component.ts
@@ -12,8 +12,8 @@ interface PiPVideoElement extends HTMLVideoElement {
         <div style="border: gray 1px solid;border-radius:5px;display:none;">
             <video #vid [src]="vidUrl" [muted]="muted" autoplay></video>
         </div>
-        <div class="pip-video">
-            <button mat-icon-button [style.display]="src ? 'block' : 'none'"
+        <div>
+            <button [style.display]="src ? 'block' : 'none'"
                 matTooltip="Show Video"
                 matTooltipPosition="left"
                 [disabled]="pipMode" (click)="initPiP()">
@@ -26,19 +26,7 @@ interface PiPVideoElement extends HTMLVideoElement {
             </button>-->                
         </div>
     `,
-    styles: [`
-        .pip-video { 
-            display:flex; 
-            max-width:85px; 
-            z-index: 200;
-            position: absolute;
-            top: 80px;
-            right: 19px;
-            background-color: white;
-            border: gray 1px solid;
-            border-radius: 5px;
-        }
-    `]
+    styles: []
 })
 export class PiPVideoComponent implements OnInit {
     
@@ -50,6 +38,7 @@ export class PiPVideoComponent implements OnInit {
     @Input() muted:boolean=true;
     @Output() resize:EventEmitter<any>= new EventEmitter();
     @Output() change:EventEmitter<any>= new EventEmitter();
+    @Output() click:EventEmitter<any>= new EventEmitter();
     @ViewChild('vid', {static: true}) vid: ElementRef;
 
     constructor() {}
@@ -92,6 +81,7 @@ export class PiPVideoComponent implements OnInit {
             await this.pipVideo.requestPictureInPicture(); 
             this.pipMode= true;     
             this.vid.nativeElement.play();
+            this.click.emit(true);
         }
         catch(e) {
             this.pipMode= false;