+++ src/app/modules/map/nextpoint.component.ts
@@ -12,29 +12,39 @@ total: number - total number of points
     selector: 'route-nextpoint',
     changeDetection: ChangeDetectionStrategy.OnPush,
 	template: `
-        <div>
+        <div class="mat-app-background">
             <div style="display: flex; flex-wrap: nowrap;" >
                 <div>
-                    <button [disabled]="index==0"
+                    <button mat-stroked-button class="nav-button"
+                        color="primary"
+                        matTooltip="Previous point"
+                        matTootipPosition="top"
+                        [disabled]="index==0"
                         (click)="changeIndex(-1)">
-                        <mat-icon>navigate_before</mat-icon><br>
-                        <span>Prev</span>			
+                        <mat-icon>skip_previous</mat-icon><br>		
                     </button>
                 </div>
                 <div>                          
-                    <button [disabled]="index>=total-1"
+                    <button mat-stroked-button class="nav-button"
+                        color="primary"
+                        matTooltip="Next point"
+                        matTootipPosition="top"
+                        [disabled]="index>=total-1"
                         (click)="changeIndex(1)">
-                        <mat-icon>navigate_next</mat-icon><br>
-                        <span>Next</span>			
+                        <mat-icon>skip_next</mat-icon><br>		
                     </button>                         
                 </div>            
             </div>   
-            <div style="background-color: rgb(221,221,221);font-family:roboto;"> 
+            <div style="font-family:roboto;height: 32px;line-height: 33px;"> 
                 {{index+1}} of {{total}}
             </div>      
         </div>
 	`,
-    styleUrls: []
+    styles: [`
+        .nav-button {
+            height:40px;
+        }
+    `]
 })
 export class RouteNextPointComponent {
     @Input() index: number;